Module velvet::prelude

Structs

Berendsen
Berendsen weak coupling thermostat.

Boltzmann
Maxwell Boltzmann style velocity distribution.

Cell
Matrix representation of a 3D bounding box.

Configuration
High-level configuration options.

ConfigurationBuilder
Constructor for the Configuration type.

Forces
Force acting on each atom in the system.

Harmonic
Harmonic style pair potential.

KineticEnergy
Kinetic energy of the whole system

LennardJones
Lennard-Jones style pair potential.

Mie
Mie style pair potential.

MolecularDynamics

Morse
Morse style pair potential.

NoseHoover
Nose-Hoover style thermostat.

NullThermostat
Placeholder thermostat algorithm which applies no temperature controls.

PairEnergy
Potential energy due to pairwise potentials.

PairForces
Force acting on each atom in the system due to pairwise potentials.

PotentialEnergy
Potential energy of the whole system.

Potentials
Container type to hold instances of each potential in the system.

PotentialsBuilder
Constructor for the Potentials type.

Simulation

Specie

System
Collection of atomic properties and bonding information.

Temperature
Instantaneous temperature of the system.

TotalEnergy
Total energy of the system.

VelocityVerlet
Velocity Verlet integration algorithm.

Enums

Element
Every element on the periodic table.

Traits

Integrator
A numerical integration algorithm.

IntrinsicProperty
Calculates a system-wide property without using the applied potentials.

Output

PairPotential
Shared behavior for pair potentials.

Propagator

Property
Calculates a system-wide property.

Thermostat
An algorithm used to control simulation temperature.

VelocityDistribution
Shared behavior for algorithms that can initialize a velocity distribution.

Functions

import_poscar
Constructs a System from POSCAR data.